--- src/main/java/Backend/sculptor/Config/WebConfig.java
@@ -1,20 +1,0 @@
-package Backend.sculptor.Config;
-
-import Backend.sculptor.OAuth.Annotation.CurrentUserArgumentResolver;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-    private final CurrentUserArgumentResolver currentUserArgumentResolver;
-
-    @Override
-    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> argumentResolvers) {
-        argumentResolvers.add(currentUserArgumentResolver);
-    }
-}
