--- src/main/java/Backend/sculptor/OAuth/Annotation/CurrentUserArgumentResolver.java
@@ -1,32 +1,0 @@
-package Backend.sculptor.OAuth.Annotation;
-
-import Backend.sculptor.User.Entity.SessionUser;
-import jakarta.servlet.http.HttpSession;
-import lombok.RequiredArgsConstructor;
-import org.springframework.core.MethodParameter;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.support.WebDataBinderFactory;
-import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.method.support.ModelAndViewContainer;
-
-@RequiredArgsConstructor
-@Component
-public class CurrentUserArgumentResolver implements HandlerMethodArgumentResolver {
-    private final HttpSession httpSession;
-
-    @Override
-    public boolean supportsParameter(MethodParameter parameter) {
-        boolean isLoginUserAnnotation = parameter.getParameterAnnotation(CurrentUser.class) != null;
-        boolean isUserClass = SessionUser.class.equals(parameter.getParameterType());
-        return isLoginUserAnnotation && isUserClass;
-    }
-
-    @Override
-    public Object resolveArgument(MethodParameter parameter,
-                                  ModelAndViewContainer mavContainer,
-                                  NativeWebRequest webRequest,
-                                  WebDataBinderFactory binderFactory) throws Exception {
-        return httpSession.getAttribute("user");
-    }
-}
