--- src/main/java/Backend/sculptor/OAuth/Controller/OAuthController.java
@@ -1,28 +1,0 @@
-package Backend.sculptor.OAuth.Controller;
-
-import Backend.sculptor.OAuth.Annotation.CurrentUser;
-import Backend.sculptor.OAuth.PrincipalDetails;
-import Backend.sculptor.User.Entity.SessionUser;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class OAuthController {
-    @GetMapping("/loginForm")
-    public String home() {
-        return "loginForm";
-    }
-
-    @GetMapping("/private")
-    public String privatePage() {
-        return "privatePage";
-    }
-
-    @GetMapping("/admin")
-    public String adminPage() {
-        return "adminPage";
-    }
-}
