--- src/main/java/Backend/sculptor/OAuth/Handler/CustomAuthenticationSuccessHandler.java
@@ -1,19 +1,0 @@
-package Backend.sculptor.OAuth.Handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import jakarta.servlet.http.HttpSession;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class CustomAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        System.out.println("Session ID = " + request.getSession().getId());
-        response.sendRedirect("/");
-    }
-}
