--- src/main/java/Backend/sculptor/OAuth/Handler/CustomLogoutSuccessHandler.java
@@ -1,21 +1,0 @@
-package Backend.sculptor.OAuth.Handler;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class CustomLogoutSuccessHandler implements LogoutSuccessHandler {
-    @Override
-    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType("application/json");
-        response.getWriter().print("{\"code\": 200, \"message\": \"Logout Success\", \"data\": null}");
-        response.sendRedirect("/loginForm");
-        response.getWriter().flush();
-    }
-}
