--- src/main/java/Backend/sculptor/OAuth/MemberInfo/OAuth2MemberInfo.java
@@ -1,9 +1,0 @@
-package Backend.sculptor.OAuth.MemberInfo;
-
-public interface OAuth2MemberInfo {
-    String getProviderId();
-    String getProvider();
-    String getName();
-    String getEmail();
-    String getProfileImage();
-}
