--- src/main/java/Backend/sculptor/Stone/Entity/AchieveStatus.java
@@ -1,11 +1,0 @@
-package Backend.sculptor.Stone.Entity;
-
-public enum AchieveStatus {
-    A, B, C
-
-}
-/*
-A: 달성
-B: 절반 달성
-C: 미달성
- */
