--- src/main/java/Backend/sculptor/Stone/Entity/Category.java
@@ -1,5 +1,0 @@
-package Backend.sculptor.Stone.Entity;
-
-public enum Category {
-    WORKOUT, STUDY, ETC
-}
