--- src/main/java/Backend/sculptor/Stone/Entity/Item.java
@@ -1,22 +1,0 @@
-package Backend.sculptor.Stone.Entity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class Item {
-    @Id @GeneratedValue
-    @Column(name = "item_id")
-    private UUID id;
-
-    private String itemName;
-    private int itemPrice;
-}
