--- src/main/java/Backend/sculptor/User/Entity/SessionUser.java
@@ -1,15 +1,0 @@
-package Backend.sculptor.User.Entity;
-
-
-import lombok.Getter;
-
-@Getter
-public class SessionUser {
-    private String name;
-    private String profileImage;
-
-    public SessionUser(Users user) {
-        this.name = user.getName();
-        this.profileImage = user.getProfileImage();
-    }
-}
