+++ src/main/java/backend/sculptor/domain/SculptorApplication.java
@@ -1,0 +1,19 @@
+package backend.sculptor.domain;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
+@SpringBootApplication
+public class SculptorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SculptorApplication.class, args);
+	}
+
+	@Bean
+	public BCryptPasswordEncoder bCryptPasswordEncoder() {
+		return new BCryptPasswordEncoder();
+	}
+}
