+++ src/main/java/backend/sculptor/domain/stone/entity/AchieveStatus.java
@@ -1,0 +1,11 @@
+package backend.sculptor.domain.stone.entity;
+
+public enum AchieveStatus {
+    A, B, C
+
+}
+/*
+A: 달성
+B: 절반 달성
+C: 미달성
+ */
