+++ src/main/java/backend/sculptor/domain/stone/entity/Category.java
@@ -1,0 +1,5 @@
+package backend.sculptor.domain.stone.entity;
+
+public enum Category {
+    WORKOUT, STUDY, ETC
+}
