+++ src/main/java/backend/sculptor/domain/stone/entity/StoneStatus.java
@@ -1,0 +1,12 @@
+package backend.sculptor.domain.stone.entity;
+
+public enum StoneStatus {
+    MOSS, PLANT, S_CRACK, L_CRACK, BROKEN
+}
+/*
+3일 MOSS : 이끼
+5일 PLANT : 식물
+7일 S_CRACK : 실금 (small_crack)
+10일 L_CRACK : 균열 (large_crack)
+12일 BROKEN : 부서짐
+ */
