+++ src/main/java/backend/sculptor/domain/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package backend.sculptor.domain.user.repository;
+
+import backend.sculptor.domain.user.entity.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<Users,Long> {
+    public Optional<Users> findByName(String name);
+}
