+++ src/main/java/backend/sculptor/global/auth/controller/OAuthController.java
@@ -1,0 +1,22 @@
+package backend.sculptor.global.auth.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class OAuthController {
+    @GetMapping("/loginForm")
+    public String home() {
+        return "loginForm";
+    }
+
+    @GetMapping("/private")
+    public String privatePage() {
+        return "privatePage";
+    }
+
+    @GetMapping("/admin")
+    public String adminPage() {
+        return "adminPage";
+    }
+}
