+++ src/main/java/backend/sculptor/global/auth/handler/CustomAuthenticationSuccessHandler.java
@@ -1,0 +1,18 @@
+package backend.sculptor.global.auth.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+public class CustomAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
+    @Override
+    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        System.out.println("Session ID = " + request.getSession().getId());
+        response.sendRedirect("/");
+    }
+}
