+++ src/main/java/backend/sculptor/global/auth/handler/CustomLogoutSuccessHandler.java
@@ -1,0 +1,21 @@
+package backend.sculptor.global.auth.handler;
+
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.web.authentication.logout.LogoutSuccessHandler;
+import org.springframework.stereotype.Component;
+
+import java.io.IOException;
+
+@Component
+public class CustomLogoutSuccessHandler implements LogoutSuccessHandler {
+    @Override
+    public void onLogoutSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
+        response.setStatus(HttpServletResponse.SC_OK);
+        response.setContentType("application/json");
+        response.getWriter().print("{\"code\": 200, \"message\": \"Logout Success\", \"data\": null}");
+        response.sendRedirect("/loginForm");
+        response.getWriter().flush();
+    }
+}
