+++ src/main/java/backend/sculptor/global/auth/memberInfo/OAuth2MemberInfo.java
@@ -1,0 +1,9 @@
+package backend.sculptor.global.auth.memberInfo;
+
+public interface OAuth2MemberInfo {
+    String getProviderId();
+    String getProvider();
+    String getName();
+    String getEmail();
+    String getProfileImage();
+}
