--- src/test/java/Backend/sculptor/SculptorApplicationTests.java
@@ -1,13 +1,0 @@
-package Backend.sculptor;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SculptorApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
