+++ src/test/java/backend/sculptor/domain/SculptorApplicationTests.java
@@ -1,0 +1,13 @@
+package backend.domain;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SculptorApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
